+++ doc/note.md
@@ -2,8 +2,7 @@
 
 ## Difference between `fact` and `predicate`
 
-- Fact and Predicate are different in the database. At the top level, the facts
-satisfying each predicate are represented by a list of structures as below
+- Fact and Predicate are different in the database. At the top level, the facts satisfying each predicate are represented by a list of structures as below
 
 Data Structures
 
@@ -36,3 +35,56 @@ Data Structures
 - simtri(contri)(A,B,C,E,F,G)
 - circle(O,A,B,C)
 - cyclic(A,B,C,D)
+
+
+## Database Implementation Details
+
+1. The database should be managed incrementally. The operations we need are
+    - DB(with existing facts) + new facts -> New DB(with fixed point facts)
+
+DB structure
+
+- lines
+- congs
+- mid facts
+- eqangle / eqratio facts
+- para facts
+- perp facts
+- circle facts
+
+
+The process of reaching the fixed point
+
+- Input: DB with existing facts, A list of new facts waited to be added
+- Output: New DB with updated facts, which has reached the fixed points
+
+
+Algorithm 1 (DB, facts_to_add) -> new DB
+
+
+WHILE facts_to_add NOT EMPTY
+
+    fact = facts_to_add.pop(0)
+
+    # ? fact could be used before
+    # if the fact was used and the database is the same
+    # then no need to update
+    # we need to record the version of the database
+
+    IF fact IN used AND used[fact] == DB.version
+        CONTINUE
+
+    all_pforms = DB.all_pforms(fact)
+
+    new_facts = { all_rules(p) for p in all_pforms }
+
+    FOR new_fact IN new_facts
+
+        IF DB.contains(new_fact) CONTINUE
+
+        facts_to_add.append(new_fact)
+
+    IF NOT DB.contains(fact)
+        DB.add(fact)           # all newly added facts should be here
+        DB.version_update()
+
